Tests/LibWeb/Text/expected/HTML/HTMLMediaElement-adoption-round-trip-during-fetch.txt

video survived adoption round-trip during fetch
video error event (error=4)
